cleanup reused methods and made base Controllers for navigation controller and collection view controller
