+++ MemeIt/Controllers/FavoritesCollectionViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class FavoritesCollectionViewController: UICollectionViewController, ViewControllerMemeController {
+class FavoritesCollectionViewController: BaseCollectionViewController, ViewControllerMemeController {
     
     // MARK: - Properties
     
@@ -172,19 +172,6 @@ class FavoritesCollectionViewController: UICollectionViewController, ViewControl
             return "Add to favorites"
         }
     }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout
-
-extension FavoritesCollectionViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let viewWidth = collectionView.frame.width - 8
-        let cellWidth = viewWidth / 3 - 2
-        let size = CGSize(width: cellWidth, height: cellWidth - 2)
-        return size
-        
-    }
-    
 }
 
 // MARK: - UITabelViewDelegate / UITableViewDataSource
