+++ MemeIt/Views/CreateMemeDetailViewController.storyboard
@@ -98,10 +98,10 @@
             </objects>
             <point key="canvasLocation" x="1865" y="114"/>
         </scene>
-        <!--Navigation Controller-->
+        <!--Base Navigation Controller-->
         <scene sceneID="J4R-Y7-QOH">
             <objects>
-                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="JiS-0E-hfU" sceneMemberID="viewController">
+                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="JiS-0E-hfU" customClass="BaseNavigationController" customModule="MemeIt" customModuleProvider="target" sceneMemberID="viewController">
                     <toolbarItems/>
                     <navigationBar key="navigationBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="CYw-ef-f8T">
                         <rect key="frame" x="0.0" y="44" width="414" height="44"/>
